Develop

+++ backend/src/controllers/member.ts
@@ -1,4 +1,6 @@
 import HttpError from "../lib/exceptions/error-handler";
+import { validateSchemaAsync } from "../lib/validators";
+import { inviteSchema } from "../lib/validators/validate-invite";
 import Member from "../models/Member";
 import { InviteController } from "./invite";
 
@@ -12,6 +14,9 @@ export class MemberController {
   ];
 
   public async create(serverId: number, userId: number, inviteCode: string) {
+    await validateSchemaAsync(inviteSchema, {
+      inviteCode,
+    });
     const isAlreadyMember = await this.isAlreadyMember(serverId, userId);
     if (isAlreadyMember) {
       throw new HttpError("User is already a member", 409);
@@ -23,10 +28,12 @@ export class MemberController {
       const newMember = await Member.query().insertAndFetch({
         member_id: userId,
         server_id: serverId,
+        is_owner: true,
       });
       return {
         id: newMember.id,
         serverId: newMember.server_id,
+        userId: newMember.member_id,
       };
     }
   }
@@ -55,12 +62,26 @@ export class MemberController {
   }
 
   public async delete(serverId: number, userId: number) {
+    const isOwner = await this.isOwner(serverId, userId);
+    if (!isOwner) {
+      throw new HttpError("Not the owner", 401);
+    }
     return await Member.query()
       .where({
         member_id: userId,
         server_id: serverId,
       })
       .delete();
+  }
+
+  private async isOwner(serverId: number, userId: number) {
+    return Boolean(
+      Member.query().where({
+        member_id: userId,
+        server_id: serverId,
+        is_owner: true,
+      }),
+    );
   }
 
   private async isAlreadyMember(serverId: number, userId: number) {
